review/requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.5
